added LICENSE
added MPL 2.0.  I am considering reverting to the MIT license, but for now I'd rather if anybody uses this code to include a copy of the source in the project.  This way I can guarantee the only software written with this open-source engine is also open-source for the betterment of public knowledge.